AWS Lambda Node update
Lambda is ending support for Node 6.10, so we need to update to 8.10.  With that version on my machine, I'm also running `npm audit` to update all known vulnerabilities.  All of the tests pass successfully, so this change will be pushed via `npm run deploy`, and the AWS Lambda will be updated.